+++ Items/WildernessGuide/WildernessGuide.cs
@@ -1,0 +1,50 @@
+﻿using System.Collections.Generic;
+using Terraria;
+using Terraria.ID;
+using Terraria.ModLoader;
+
+namespace Combinations.Items.WildernessGuide
+{
+    public class WildernessGuide : CombinationsBaseModItem
+    {
+        public override void SetStaticDefaults()
+        {
+            Tooltip.SetDefault("Allows the collection of Vine Rope from vines\n" +
+                "Prevents you from hurting critters while in the inventory\n" +
+                "'Does not explain the forces of nature'");
+        }
+
+        public override void SetDefaults()
+        {
+            Item.accessory = true;
+            Item.width = 20;
+            Item.height = 20;
+            Item.value = Item.sellPrice(0, 2);
+            Item.rare = ItemRarityID.Blue;
+            Item.stack = 1;
+        }
+
+        public override void AddRecipes()
+        {
+            Recipe recipe = CreateRecipe();
+            recipe.AddIngredient(ItemID.CordageGuide);
+            recipe.AddIngredient(ItemID.DontHurtCrittersBook);
+            recipe.AddTile(TileID.TinkerersWorkbench);
+            recipe.Register();
+        }
+
+        public override void UpdateAccessory(Player player, bool hideVisual)
+        {
+            player.cordage = true;
+            player.dontHurtCritters = true;
+        }
+
+        public static int ItemType() => ModContent.ItemType<WildernessGuide>();
+
+        public override List<int> IncompatibleAccessories() =>
+            new List<int>()
+            {
+                ItemType(),
+            };
+    }
+}